+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## `master`
 
+## `1.1.0`
+
 - New: #34 support `google.protobuf.Duration`
 - New: #26 support [`google.protobuf.Any`](https://developers.google.com/protocol-buffers/docs/proto3#any)
 - New: #56 provide public API for reading and writing `VarInt`s